Enforce the certificate to use in `set_up_signing`
That way when we temporarily switch that lane to use `readonly: false` when we need to regenerate the profile, this makes sure that the newly generated profile uses that specific cert, instead of _fastlane_ trying to guess and potentially picking your _personal_ Apple Development profile.

This is what happened to me when I renewed the cert 2 days ago trying to fix code signing issues when developing `hostmgr` locally.