RFC specifies ETags are expected to contain quotes
Despite this looking odd at first, the literal quotes are supposed to be fully part of the ETag string; in fact, the spec says that an ETag might be prefixed with `W/` before the `quoted-string`, e.g. `W/"a1b2c3"`, so it'd be invalid to remove quotes from that.

This is specified explicitly in https://www.rfc-editor.org/rfc/rfc2616#section-3.11:
```
      entity-tag = [ weak ] opaque-tag
      weak       = "W/"
      opaque-tag = quoted-string

      quoted-string  = ( <"> *(qdtext | quoted-pair ) <"> )
```

So better respect this and keep the literal quotes for consistency with the spec and what the S3 API returns (and expects back)